third_party/WebKit/LayoutTests/external/wpt/XMLHttpRequest/open-url-encoding-expected.txt

This is a testharness.js-based test.
PASS percent encode characters
FAIL lone surrogate assert_equals: expected "&%2365533;" but got "%26%2355357%3B"
Harness: the test ran to completion.
